django/requirements.txt
asgiref==3.5.0
backoff==1.10.0
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.12
Deprecated==1.2.13
Django==4.0.4
googleapis-common-protos==1.56.0
grpcio==1.44.0
idna==3.3
opentelemetry-api==1.10.0
opentelemetry-exporter-otlp==1.10.0
opentelemetry-exporter-otlp-proto-grpc==1.10.0
opentelemetry-exporter-otlp-proto-http==1.10.0
opentelemetry-instrumentation==0.29b0
opentelemetry-instrumentation-django==0.29b0
opentelemetry-instrumentation-wsgi==0.29b0
opentelemetry-proto==1.10.0
opentelemetry-sdk==1.10.0
opentelemetry-semantic-conventions==0.29b0
opentelemetry-util-http==0.29b0
protobuf==3.20.0
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
wrapt==1.14.0
